backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.2
keras==3.13.2
tensorflow-cpu==2.18.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==11.3.0
gunicorn==23.0.0
